fix(ios): report non-finite progress instead of trapping
Introduced by the Swift port of the progress config in 4be3a03 and not
present in any released build, so this is a regression against the
Objective-C rather than a bug anyone has hit.

The count throttle compared `Int(floor(progress * count)) >= tick`.
Int(Float) traps on a non-finite value, and +infinity reaches that line: a
caller computing written/total with a total of 0 and a non-zero written
passes infinity, which satisfies the `> 0` guard just above. NaN cannot
reach it, because NaN fails every comparison including `> 0`.

The Objective-C this replaced did `(int)floorf(...)`, and converting a
non-finite float to int saturates on arm64 - INT_MAX, so the comparison was
true and the event was reported. Measured rather than assumed:

  (int)floorf(inf*10) = 2147483647  (INT_MAX = 2147483647)
  >= tick(1) : YES

Comparing in floating point keeps that. `floor(progress * count) >=
Float(tick)` is true for +infinity and identical for every finite value -
checked across the whole 0...1 progress range against counts of 1, 2, 3, 5,
10, 100 and 1000 and ticks 1 through 12, with no disagreement.

Whether an iOS call site can actually produce a zero total with a non-zero
written - an upload whose expected size reports 0, a response carrying
Content-Length: 0 that still delivers bytes - is exactly what this avoids
having to prove.

The test fails without the fix, and not as an assertion: it takes the whole
run down.

  Swift/arm64-apple-ios-simulator.swiftinterface:39402: Fatal error: Float
  value cannot be converted to Int because it is either infinite or NaN
  Restarting after unexpected exit, crash, or test timeout; summary will
  include totals from previous launches.

A NaN case is covered too, so the `> 0` guard that keeps NaN away from the
conversion cannot be removed without something failing.

+++ ios/ReactNativeBlobUtilProgress.swift
@@ -47,7 +47,13 @@ public class ReactNativeBlobUtilProgress: NSObject {
         var withinCount = true
         let countValue = count.floatValue
         if countValue > 0 && nextProgress.floatValue > 0 {
-            withinCount = Int(floor(nextProgress.floatValue * countValue)) >= tick
+            // Compared in floating point rather than through Int(). A caller
+            // dividing written by a total of 0 passes +infinity, which is > 0 and
+            // so reaches here, and Int(Float) traps on a non-finite value. The
+            // Objective-C this replaced saturated the conversion instead and
+            // reported the event, which is what this keeps doing. Identical for
+            // every finite value.
+            withinCount = floor(nextProgress.floatValue * countValue) >= Float(tick)
         }
 
         let timeStamp = Date().timeIntervalSince1970
